+++ src/components/Category/CategoryList.tsx
@@ -33,4 +33,4 @@ function CategoryList() {
     </ActionableLayout>
 }
 
-export default React.memo(CategoryList);
+export default CategoryList;
