+++ src/components/Layout/BurgerMenu.css
@@ -15,3 +15,7 @@
   transform-origin: 1px;
   transition: all 0.3s linear;
 }
+
+.BurgerMenu--hidden {
+  display: none;
+}
