+++ src/components/Layout/BurgerMenuContent.css
@@ -1,0 +1,4 @@
+.BurgerMenuContent.Animated {
+  position: absolute;
+  width: 100%;
+}
